requirements-dev.txt
aiohappyeyeballs==2.4.0
aiohttp==3.11.16
aiohttp-retry==2.8.3
aiosignal==1.3.1
amqp==5.3.1
annotated-types==0.7.0
anyio==4.7.0
appdirs==1.4.4
argcomplete==3.5.3
asgiref==3.8.1
asn1crypto==1.5.1
asttokens==3.0.0
asyncpg==0.30.0
attrs==24.2.0
auth0-python==4.9.0
backoff==2.2.1
bcrypt==4.2.1
beautifulsoup4==4.12.3
billiard==4.2.1
blinker==1.9.0
bs4==0.0.2
build==1.2.2.post1
cachetools==5.5.0
celery==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
chroma-hnswlib==0.7.6
chromadb==0.5.23
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cmdstanpy==1.2.5
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.1
crewai==0.108.0
cryptography==43.0.3
cycler==0.12.1
Cython==3.0.11
dataclasses-json==0.6.7
decorator==5.2.1
Deprecated==1.2.15
deprecation==2.1.0
distro==1.9.0
docopt==0.4.0
docstring_parser==0.16
durationpy==0.9
et_xmlfile==2.0.0
executing==2.2.0
faker>=25.0
fastapi==0.115.6
filelock==3.16.1
Flask==3.1.0
flatbuffers==24.3.25
fonttools==4.54.1
frozenlist==1.4.1
fsspec==2025.2.0
git-filter-repo==2.47.0
google-auth==2.37.0
googleapis-common-protos==1.66.0
greenlet==3.1.1
grpcio==1.70.0
gunicorn==23.0.0
h11==0.14.0
h2==4.3.0
holidays==0.62
hpack==4.1.0
httpcore==1.0.5
httptools==0.6.4
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.26.5
humanfriendly==10.0
hyperframe==6.1.0
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
instructor==1.7.9
ipython==9.0.2
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.4
jiter==0.8.2
json5==0.12.0
json_repair==0.40.0
jsonpatch==1.33
jsonpickle==4.0.5
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.7
kombu==5.5.2
kubernetes==31.0.0
langchain==0.3.19
langchain-chroma==0.1.4
langchain-community==0.3.18
langchain-core==0.3.39
langchain-openai==0.3.7
langchain-text-splitters==0.3.6
langgraph>=0.2
langsmith==0.3.11
litellm==1.60.2
mandrill==1.0.60
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.22.0
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mmh3==5.0.1
monotonic==1.6
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
networkx==3.4.2
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.20.1
openai==1.64.0
openpyxl==3.1.5
opentelemetry-api==1.31.1
opentelemetry-exporter-otlp-proto-common==1.31.1
opentelemetry-exporter-otlp-proto-grpc==1.31.1
opentelemetry-exporter-otlp-proto-http==1.31.1
opentelemetry-instrumentation==0.52b1
opentelemetry-instrumentation-asgi==0.52b1
opentelemetry-instrumentation-fastapi==0.52b1
opentelemetry-proto==1.31.1
opentelemetry-sdk==1.31.1
opentelemetry-semantic-conventions==0.52b1
opentelemetry-util-http==0.52b1
orjson==3.10.7
overrides==7.7.0
packaging==24.2
pandas==2.2.3
parso==0.8.4
pdfminer.six==20250327
pdfplumber==0.11.6
pika==1.3.2
pillow==11.1.0
pipx==1.7.1
platformdirs==4.3.6
postgrest==1.1.1
posthog==3.7.4
prompt_toolkit==3.0.50
propcache==0.3.1
prophet==1.1.6
protobuf==5.29.3
psycopg==3.2.10
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.11.9
pydantic-settings==2.5.2
pydantic_core==2.33.2
Pygments==2.18.0
PyJWT==2.10.1
pyodbc==5.2.0
pyOpenSSL==24.2.1
pyparsing==3.2.0
pypdfium2==4.30.1
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
pytesseract==0.3.13
pytest>=8.2
pytest-asyncio>=0.23
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pyvis==0.3.2
PyYAML==6.0.2
RapidFuzz==3.10.0
realtime==2.7.0
redis==5.2.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.24.0
rsa==4.9
setuptools==75.6.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
snowflake-connector-python==3.12.3
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.35
stack-data==0.6.3
stanio==0.5.1
starlette==0.41.3
storage3==0.12.1
StrEnum==0.4.15
supabase==2.18.1
supabase_auth==2.12.3
supabase_functions==0.10.1
sympy==1.13.3
tenacity==9.1.2
tiktoken==0.9.0
tokenizers==0.20.3
tomli==2.2.1
tomli_w==1.2.0
tomlkit==0.13.2
tqdm==4.67.1
traitlets==5.14.3
twilio==9.4.0
typer==0.15.1
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.2.3
userpath==1.9.2
uv==0.6.12
uvicorn==0.32.1
vine==5.1.0
waitress==3.0.2
watchfiles==1.0.3
wcwidth==0.2.13
websocket-client==1.8.0
websockets==14.1
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.0
yarl==1.19.0
zipp==3.21.0
zstandard==0.23.0
